remove src path
