+++ src/streamed/index.ts
@@ -14,12 +14,12 @@
  */
 
 import Web3 from "web3";
-import { BlockNumber } from "web3-core";
+import { BlockNumber, TransactionReceipt } from "web3-core";
+import { Contract } from "web3-eth-contract";
 import {
   BatchExchange,
   BatchExchangeArtifact,
   ContractArtifact,
-  deployment,
 } from "../contracts";
 import { IndexedOrder } from "../encoding";
 import { AnyEvent } from "./events";
@@ -302,4 +302,27 @@ export class InvalidAuctionStateError extends Error {
   constructor(public readonly block: number, public readonly inner: Error) {
     super(`invalid auction state at block ${block}: ${inner.message}`);
   }
+}
+
+/**
+ * Get a contract deployment, returning both the web3 contract object as well as
+ * the transaction receipt for the contract deployment.
+ *
+ * @throws If the contract is not deployed on the network the web3 provider is
+ * connected to.
+ */
+export async function deployment<C extends Contract>(
+  web3: Web3,
+  { abi, networks }: ContractArtifact,
+): Promise<[C, TransactionReceipt]> {
+  const chainId = await web3.eth.getChainId();
+  const network = networks[chainId];
+  if (!networks) {
+    throw new Error(`not deployed on network with chain ID ${chainId}`);
+  }
+
+  const tx = await web3.eth.getTransactionReceipt(network.transactionHash);
+  const contract = new web3.eth.Contract(abi, network.address);
+
+  return [contract as C, tx];
 }
