+++ test/models/streamed/state.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
-import { BatchExchange, DEFAULT_ORDERBOOK_OPTIONS } from "@src/index";
-import { AuctionState } from "@src/streamed/state";
-import { AnyEvent, EventName, EventValues } from "@src/streamed/events";
+import { BatchExchange, DEFAULT_ORDERBOOK_OPTIONS } from "../../../src";
+import { AuctionState } from "../../../src/streamed/state";
+import { AnyEvent, EventName, EventValues } from "../../../src/streamed/events";
 
 function auctionState(): AuctionState {
   return new AuctionState({ ...DEFAULT_ORDERBOOK_OPTIONS, strict: true });
